Format with yapf and limit line length to 80 char
yapf -irp --style=pep8 .

Some hand-crafted corrections added

+++ sandbox/rocky/tf/core/parameterized.py
@@ -4,8 +4,8 @@ from rllab.core import Serializable
 from rllab.misc.tensor_utils import flatten_tensors, unflatten_tensors
 import tensorflow as tf
 
-
 load_params = True
+
 
 @contextmanager
 def suppress_params_loading():
@@ -44,7 +44,9 @@ class Parameterized(object):
         if tag_tuple not in self._cached_param_dtypes:
             params = self.get_params(**tags)
             param_values = tf.get_default_session().run(params)
-            self._cached_param_dtypes[tag_tuple] = [val.dtype for val in param_values]
+            self._cached_param_dtypes[tag_tuple] = [
+                val.dtype for val in param_values
+            ]
         return self._cached_param_dtypes[tag_tuple]
 
     def get_param_shapes(self, **tags):
@@ -52,7 +54,9 @@ class Parameterized(object):
         if tag_tuple not in self._cached_param_shapes:
             params = self.get_params(**tags)
             param_values = tf.get_default_session().run(params)
-            self._cached_param_shapes[tag_tuple] = [val.shape for val in param_values]
+            self._cached_param_shapes[tag_tuple] = [
+                val.shape for val in param_values
+            ]
         return self._cached_param_shapes[tag_tuple]
 
     def get_param_values(self, **tags):
@@ -62,27 +66,29 @@ class Parameterized(object):
 
     def set_param_values(self, flattened_params, **tags):
         debug = tags.pop("debug", False)
-        param_values = unflatten_tensors(
-            flattened_params, self.get_param_shapes(**tags))
+        param_values = unflatten_tensors(flattened_params,
+                                         self.get_param_shapes(**tags))
         ops = []
         feed_dict = dict()
         for param, dtype, value in zip(
-                self.get_params(**tags),
-                self.get_param_dtypes(**tags),
+                self.get_params(**tags), self.get_param_dtypes(**tags),
                 param_values):
             if param not in self._cached_assign_ops:
-                assign_placeholder = tf.placeholder(dtype=param.dtype.base_dtype)
+                assign_placeholder = tf.placeholder(
+                    dtype=param.dtype.base_dtype)
                 assign_op = tf.assign(param, assign_placeholder)
                 self._cached_assign_ops[param] = assign_op
                 self._cached_assign_placeholders[param] = assign_placeholder
             ops.append(self._cached_assign_ops[param])
-            feed_dict[self._cached_assign_placeholders[param]] = value.astype(dtype)
+            feed_dict[self._cached_assign_placeholders[param]] = value.astype(
+                dtype)
             if debug:
                 print("setting value of %s" % param.name)
         tf.get_default_session().run(ops, feed_dict=feed_dict)
 
     def flat_to_params(self, flattened_params, **tags):
-        return unflatten_tensors(flattened_params, self.get_param_shapes(**tags))
+        return unflatten_tensors(flattened_params,
+                                 self.get_param_shapes(**tags))
 
     def __getstate__(self):
         d = Serializable.__getstate__(self)
@@ -95,7 +101,8 @@ class Parameterized(object):
         Serializable.__setstate__(self, d)
         global load_params
         if load_params:
-            tf.get_default_session().run(tf.variables_initializer(self.get_params()))
+            tf.get_default_session().run(
+                tf.variables_initializer(self.get_params()))
             self.set_param_values(d["params"])
 
 
@@ -105,6 +112,9 @@ class JointParameterized(Parameterized):
         self.components = components
 
     def get_params_internal(self, **tags):
-        params = [param for comp in self.components for param in comp.get_params_internal(**tags)]
+        params = [
+            param for comp in self.components
+            for param in comp.get_params_internal(**tags)
+        ]
         # only return unique parameters
         return sorted(set(params), key=hash)
